# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Retrieve a paginated list of customers",
                "description": "This endpoint returns a paginated list of all customers in the system.",
                "operationId": "2b23ac696cfd81c23807c2c1fd63b0b0",
                "responses": {
                    "200": {
                        "description": "Paginated list of customers retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "description": "Current page number",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "description": "List of customers",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Customer ID",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "description": "Customer's name",
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "email": {
                                                        "description": "Customer's email",
                                                        "type": "string",
                                                        "example": "johndoe@example.com"
                                                    },
                                                    "address": {
                                                        "description": "Customer's address",
                                                        "type": "string",
                                                        "example": "1234 Main St, City, Country"
                                                    },
                                                    "contact_number": {
                                                        "description": "Customer's contact number",
                                                        "type": "string",
                                                        "example": "1234567890"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "description": "URL of the first page",
                                            "type": "string",
                                            "example": "http://example.com/api/customers?page=1"
                                        },
                                        "from": {
                                            "description": "Starting item index of the current page",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "description": "Last page number",
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "last_page_url": {
                                            "description": "URL of the last page",
                                            "type": "string",
                                            "example": "http://example.com/api/customers?page=5"
                                        },
                                        "next_page_url": {
                                            "description": "URL of the next page",
                                            "type": "string",
                                            "example": "http://example.com/api/customers?page=2",
                                            "nullable": true
                                        },
                                        "path": {
                                            "description": "Base URL for pagination",
                                            "type": "string",
                                            "example": "http://example.com/api/customers"
                                        },
                                        "per_page": {
                                            "description": "Number of items per page",
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "prev_page_url": {
                                            "description": "URL of the previous page",
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "to": {
                                            "description": "Ending item index of the current page",
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "total": {
                                            "description": "Total number of items in the collection",
                                            "type": "integer",
                                            "example": 75
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customer/save": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Add or Update a Customer",
                "description": "This endpoint creates a new customer or updates an existing one based on the provided customer ID.",
                "operationId": "9264cf21dabaa7fdd6032a669d727ca4",
                "requestBody": {
                    "description": "Customer details for creation or update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Customer ID (if updating an existing customer)",
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "name": {
                                        "description": "Customer's name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "Customer's email address",
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "address": {
                                        "description": "Customer's address",
                                        "type": "string",
                                        "example": "1234 Main St, City, Country"
                                    },
                                    "contact_number": {
                                        "description": "Customer's contact number",
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created or updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Customer saved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": null
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "johndoe@example.com"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "1234 Main St, City, Country"
                                                },
                                                "contact_number": {
                                                    "type": "string",
                                                    "example": "1234567890"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-11-13T10:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-11-13T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "messages": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The name field is required."
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email must be a valid email address."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Server Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An unexpected error occurred."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}